+++ wp-content/themes/workingnyc/assets/js/search-483e095.js
@@ -1329,7 +1329,7 @@
 	  );
 
 	/**
-	 * Jobs Archive
+	 * Search results
 	 *
 	 * @author NYC Opportunity
 	 */
